requirements.txt
alabaster==0.7.12
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
autopep8==1.7.0
Babel==2.10.3
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///tmp/build/80754af9/beautifulsoup4_1631874778482/work
bleach==5.0.1
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1625814693446/work
chardet @ file:///tmp/build/80754af9/chardet_1607706768982/work
click==8.1.3
conda==4.10.3
conda-build==3.21.5
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262151086/work
cryptography @ file:///tmp/build/80754af9/cryptography_1633521537906/work
debugpy==1.6.3
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml==0.7.1
dgl-cu113==0.9.1
dglgo==0.0.2
dnspython==2.1.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.19
embeddings==0.0.8
entrypoints==0.4
fastjsonschema==2.16.2
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
gitdb==4.0.9
GitPython==3.1.29
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
huggingface-hub==0.10.1
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imagesize==1.4.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.1
ipython @ file:///tmp/build/80754af9/ipython_1632141056704/work
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
jedi @ file:///tmp/build/80754af9/jedi_1611333758854/work
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.4.3
jupyter-console==6.4.4
jupyter-core==4.11.2
jupyter-server==1.21.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
littleutils==0.2.2
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.7
notebook==6.5.1
notebook-shim==0.2.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634106693478/work
numpydoc==1.5.0
ogb==1.3.4
olefile==0.46
outdated==0.2.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
pkginfo==1.7.1
pkgutil-resolve-name==1.3.10
prometheus-client==0.15.0
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.1
psutil @ file:///tmp/build/80754af9/psutil_1612298016854/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycodestyle==2.9.1
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.10.2
Pygments==2.13.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2021.3
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
rdkit-pypi==2022.3.5
regex==2022.9.13
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016701961/work
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
sentencepiece==0.1.91
sentry-sdk==1.9.10
setproctitle==1.3.2
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1616183228191/work
sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stanza==1.1.1
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.1
toml==0.10.2
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchelastic==0.2.0
torchtext==0.11.0
torchvision==0.11.2+cu113
tornado==6.2
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==5.5.0
transformers==4.23.1
typer==0.6.1
typing-extensions==4.4.0
urllib3==1.26.12
wandb==0.13.4
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.3
yarg==0.1.9
zipp==3.9.0
